+++ apps/web/components/landing/CallToAction/index.tsx
@@ -1,11 +1,9 @@
 "use client";
 
 import Image from "next/image";
-import imageOne from "../../../assets/images/call_to_action/creator-1.jpg";
-import imageTwo from "../../../assets/images/call_to_action/creator-2.png";
-import imageThree from "../../../assets/images/call_to_action/creator-3.png";
-import imageFour from "../../../assets/images/call_to_action/creator-4.png";
+import { useTranslation } from "react-i18next";
 import kiibeeLogo from "../../../assets/images/kiibee-logo.svg";
+import { desktopCards, mobileCards } from "./utils/cards";
 import {
   Section,
   Backdrop,
@@ -24,48 +22,9 @@ import {
   CTAButton,
 } from "./styles";
 
-const orderedImages = {
-  1: imageFour.src,
-  2: imageThree.src,
-  3: imageTwo.src,
-  4: imageOne.src,
-  5: imageFour.src,
-  6: imageThree.src,
-} as const;
-
-const desktopCards = [
-  { src: orderedImages[1], left: -6.6, top: 0.8, width: 15.4, height: 31.2 },
-  { src: orderedImages[2], left: 16.0, top: 0.8, width: 13.8, height: 31.2 },
-  { src: orderedImages[3], left: 30.2, top: 0.8, width: 15.8, height: 31.2 },
-  { src: orderedImages[4], left: 46.8, top: 0.8, width: 15.7, height: 31.2 },
-  { src: orderedImages[5], left: 63.4, top: 0.8, width: 15.7, height: 31.2 },
-  { src: orderedImages[6], left: 80.0, top: 0.8, width: 13.9, height: 31.2 },
-  { src: orderedImages[5], left: -3.8, top: 32.6, width: 16.3, height: 31.8 },
-  { src: orderedImages[6], left: 12.0, top: 32.6, width: 15.5, height: 31.8 },
-  { src: orderedImages[3], left: 28.3, top: 32.6, width: 16.3, height: 31.8 },
-  { src: orderedImages[4], left: 44.9, top: 32.6, width: 15.8, height: 31.8 },
-  { src: orderedImages[1], left: 61.3, top: 32.6, width: 16.1, height: 31.8 },
-  { src: orderedImages[2], left: 77.7, top: 32.6, width: 13.8, height: 31.8 },
-  { src: orderedImages[5], left: -4.4, top: 65.0, width: 15.6, height: 34.0 },
-  { src: orderedImages[6], left: 16.0, top: 65.0, width: 14.2, height: 34.0 },
-  { src: orderedImages[3], left: 32.0, top: 65.0, width: 16.0, height: 34.0 },
-  { src: orderedImages[4], left: 48.6, top: 65.0, width: 15.8, height: 34.0 },
-  { src: orderedImages[1], left: 65.0, top: 65.0, width: 16.0, height: 34.0 },
-  { src: orderedImages[2], left: 81.5, top: 65.0, width: 13.8, height: 34.0 },
-] as const;
-
-const mobileCards = [
-  orderedImages[1],
-  orderedImages[2],
-  orderedImages[3],
-  orderedImages[4],
-  orderedImages[5],
-  orderedImages[6],
-  orderedImages[3],
-  orderedImages[4],
-];
-
 export default function CallToAction() {
+  const { t } = useTranslation();
+
   return (
     <Section>
       <Backdrop>
@@ -77,7 +36,7 @@ export default function CallToAction() {
             $width={card.width}
             $height={card.height}
           >
-            <CardImage src={card.src} alt="creator" />
+            <CardImage src={card.src} alt={t("callToAction.creatorAlt")} />
             <CardTint />
           </Card>
         ))}
@@ -87,7 +46,7 @@ export default function CallToAction() {
         <MobileGrid>
           {mobileCards.map((src, index) => (
             <Card key={`${src}-mobile-${index}`} $mobileOnly>
-              <CardImage src={src} alt="creator" />
+              <CardImage src={src} alt={t("callToAction.creatorAlt")} />
               <CardTint />
             </Card>
           ))}
@@ -102,7 +61,7 @@ export default function CallToAction() {
           <BrandLogo>
             <Image
               src={kiibeeLogo}
-              alt="Kiibee"
+              alt={t("callToAction.logoAlt")}
               fill
               priority
               style={{ objectFit: "contain" }}
@@ -110,16 +69,12 @@ export default function CallToAction() {
           </BrandLogo>
         </Brand>
         <Heading>
-          Are you a creator?
+          {t("callToAction.titleLine1")}
           <br />
-          Join Kiibee today
+          {t("callToAction.titleLine2")}
         </Heading>
-        <Subtitle>
-          Share your work, build your audience, and earn directly from your
-          content. Kiibee helps you publish, promote, and sell in one simple
-          platform.
-        </Subtitle>
-        <CTAButton type="button">Start creating</CTAButton>
+        <Subtitle>{t("callToAction.subtitle")}</Subtitle>
+        <CTAButton type="button">{t("callToAction.cta")}</CTAButton>
       </Content>
     </Section>
   );
