+++ apps/web/locals/da.json
@@ -32,5 +32,13 @@
         "description": "Stream eller download med det samme, når som helst, hvor som helst."
       }
     }
+  },
+  "callToAction": {
+    "logoAlt": "Kiibee",
+    "creatorAlt": "Skaber",
+    "titleLine1": "Er du en skaber?",
+    "titleLine2": "Bliv en del af Kiibee i dag",
+    "subtitle": "Del dit arbejde, opbyg dit publikum, og tjen direkte på dit indhold. Kiibee hjælper dig med at udgive, promovere og sælge på én enkel platform.",
+    "cta": "Begynd at skabe"
   }
 }
